+++ kipper/core/src/compiler/ast/nodes/declarations/type-declaration/class-declaration/class-member-declaration/class-constructor-declaration/class-constructor-declaration-type-semantics.ts
@@ -1,14 +1,14 @@
-import type { TypeDeclarationTypeSemantics } from "../../../type-declaration-type-semantics";
 import type { ProcessedType } from "../../../../../../../semantics";
+import type { ClassMemberDeclarationTypeSemantics } from "../class-member-declaration-type-semantics";
 
 /**
- *
+ * Type semantics for AST Node {@link ClassConstructorDeclaration}.
  * @since 0.12.0
  */
-export interface ClassConstructorDeclarationTypeSemantics extends TypeDeclarationTypeSemantics {
+export interface ClassConstructorDeclarationTypeSemantics extends ClassMemberDeclarationTypeSemantics {
 	/**
 	 * The processed type of this member property.
 	 * @since 0.12.0
 	 */
-	type: ProcessedType;
+	valueType: ProcessedType;
 }