+++ kipper/core/src/compiler/ast/nodes/declarations/type-declaration/class-declaration/class-member-declaration/class-member-declaration-semantics.ts
@@ -1,10 +1,10 @@
-import type { TypeDeclarationSemantics } from "../../type-declaration-semantics";
+import type { DeclarationSemantics } from "../../../declaration-semantics";
 
 /**
  * Semantics for AST Node {@link InterfacePropertyDeclaration}.
  * @since 0.12.0
  */
-export interface ClassMemberDeclarationSemantics extends TypeDeclarationSemantics {
+export interface ClassMemberDeclarationSemantics extends DeclarationSemantics {
 	/**
 	 * The identifier of the interface member.
 	 * @since 0.12.0