+++ kipper/core/src/compiler/ast/nodes/declarations/type-declaration/class-declaration/class-member-declaration/class-method-declaration/class-method-declaration-type-semantics.ts
@@ -1,14 +1,14 @@
-import type { TypeDeclarationTypeSemantics } from "../../../type-declaration-type-semantics";
 import type { ProcessedType } from "../../../../../../../semantics";
+import type { ClassMemberDeclarationTypeSemantics } from "../class-member-declaration-type-semantics";
 
 /**
- *
+ * Type semantics for AST Node {@link ClassConstructorDeclaration}.
  * @since 0.12.0
  */
-export interface ClassMethodDeclarationTypeSemantics extends TypeDeclarationTypeSemantics {
+export interface ClassMethodDeclarationTypeSemantics extends ClassMemberDeclarationTypeSemantics {
 	/**
-	 * The processed type of this member property.
+	 * The processed type of this member property. This always of type {@link BuiltInTypeFunc}.
 	 * @since 0.12.0
 	 */
-	type: ProcessedType;
+	valueType: ProcessedType;
 }