+++ kipper/core/src/compiler/ast/nodes/declarations/type-declaration/type-declaration-type-semantics.ts
@@ -10,5 +10,5 @@ export interface TypeDeclarationTypeSemantics extends TypeData {
 	 * The processed type of the type declaration.
 	 * @since 0.11.0
 	 */
-	type: ProcessedType;
+	valueType: ProcessedType;
 }