+++ kipper/core/src/compiler/ast/nodes/expressions/new-instantiation-expression/new-instantiation-expression-semantics.ts
@@ -1,0 +1,8 @@
+import type { ExpressionSemantics } from "../expression-semantics";
+import type { IdentifierTypeSpecifierExpression } from "../type-specifier-expression";
+import type { Expression } from "../expression";
+
+export interface NewInstantiationExpressionSemantics extends ExpressionSemantics {
+	class: IdentifierTypeSpecifierExpression;
+	args: Array<Expression>;
+}