+++ CHANGELOG.md
@@ -18,6 +18,11 @@ To use development versions of Kipper download the
 
 ### Added
 
+- Semantic checking and code generation for the `new` keyword expression to be able to create new instances of
+  classes. ([#679](https://github.com/Kipper-Lang/Kipper/issues/679))
+- Support for the typeof operator, which returns the runtime type of a value.
+  ([#663](https://github.com/Kipper-Lang/Kipper/issues/663))
+- Support for Nix Flakes and direnv, which allows for a more reproducible and consistent development environment.
 - Support for dot notation for accessing properties of objects. ([#67](https://github.com/Kipper-Lang/Kipper/issues/67))
 - Support for classes, class methods, class properties and class constructors.
   ([#665](https://github.com/Kipper-Lang/Kipper/issues/665))
@@ -45,6 +50,10 @@ To use development versions of Kipper download the
   - `semantics/runtime-internals`, which contains the runtime internal functions.
   - `semantics/types`, which contains the runtime types.
 - New classes:
+  - `NewInstantiationExpression`, which represents an AST new instantiation expression.
+  - `TypeofExpression`, which represents an AST typeof expression that returns the runtime type of an object.
+  - `TypeofTypeSpecifierExpression`, which represents an AST typeof type specifier that lets one define a type by using an object as reference
+  - `BuiltInTypeObject`, which is the base class for the compilers representation of runtime objects
   - `UniverseScope`, which represents the universe scope, where all built-in types, functions and variables are
     declared. This serves as the parent of the global scope.
   - `InterfaceDeclaration`, which represents an AST interface declaration.
@@ -101,8 +110,14 @@ To use development versions of Kipper download the
   - `CompilableType`, which represents a type that can be compiled.
   - `BuiltInReference`, which replaces the now removed type `Reference` in the `KipperProgramContext` for reference
     tracking of built-in types.
+  - `NewInstantiationExpressionSemantics`, which represents the semantics of a new instantiation expression.
+  - `NewInstantiationExpressionTypeSemantics`, which represents the type semantics of a new instantiation expression.
+  - `TypeofExpressionSemantics`, which represents the semantics of a typeof expression.
+  - `TypeofExpressionTypeSemantics`, which represents the type semantics of a typeof expression.
 - New functions:
   - `KipperTypeChecker.validArrayExpression`, which ensures that an array expression is valid.
+  - `generateInterfaceRuntimeTypeChecks` which generates runtime type checks for an interface.
+  - `getRuntimeType`, which gets the corresponding runtime representation of a built-in type.
 - New properties:
   - `BuiltInFunction.funcType`, which returns a function type for the built-in function.
   - `FunctionDeclarationTypeSemantics.type`, which returns the type of the function declaration i.e. the function type.